+++ src/acconeer/exptool/app/new/ui/main_window.py
@@ -24,7 +24,7 @@ from acconeer.exptool.app.new.app_model import AppModel
 
 from .connection_widget import ClientConnectionWidget, GenerationSelection
 from .flash_widget import FlashButton
-from .misc import ConnectionHint, ExceptionWidget, VerticalSeparator
+from .misc import ConnectionHint, ExceptionWidget, UnflashedDeviceHint, VerticalSeparator
 from .plugin_widget import PluginControlArea, PluginPlotArea, PluginSelection
 from .recording_widget import RecordingWidget
 from .utils import ScrollAreaDecorator, TopAlignDecorator
@@ -78,6 +78,7 @@ class TopBar(QFrame):
         self.layout().addWidget(VerticalSeparator(self))
         self.layout().addWidget(FlashButton(app_model, self))
         self.layout().addWidget(ConnectionHint(app_model, self))
+        self.layout().addWidget(UnflashedDeviceHint(app_model, self))
         self.layout().addStretch(1)
         self.layout().addWidget(RecordingWidget(app_model, self))